run_make_support: move `assert_recursive_eq` into `assertion_helpers`
